pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "funny_commentator"
version = "1.0.0"
description = "A funny ARK server commentator bot"
authors = [{name = "Your Name"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "discord.py",
    "python-dotenv",
    "requests",
    "rcon",
    "apscheduler",
    "tiktoken",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-asyncio",
    "pytest-mock",
    "pytest-cov",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
asyncio_mode = "auto"
